frozen-requirements.txt
icalendar==4.0.6
ruamel.yaml==0.16.12
requests==2.24.0
Flask==1.1.2
## The following requirements were added by pip freeze:
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dateutil==2.8.1
pytz==2020.1
ruamel.yaml.clib==0.2.2
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
